refactor(tui): colocate Screen::Preview into tui/screens/preview.rs
Third screen migration for #287 (Phase 2). Moved into the new
tui/screens/preview.rs: handle_key_preview, build_preview_vm,
render_preview_vm (+ its exclusive helper preview_line_spans),
preview_palette_items.

wheel_step's Screen::Diff(_) | Screen::Preview(_) | Screen::Confirm(_) => 3
arm covers three screens with one shared constant — not a Preview-exclusive
arm, so left untouched (same "shared stays" treatment as open_config/
open_help). HelpTopic::for_screen has no dedicated Preview arm either
(covered by the existing _ => HelpTopic::List wildcard) — nothing to move
there.

file_ext (view_model.rs) is shared by build_diff_vm and build_preview_vm,
so it stays in view_model.rs, widened from private to pub(crate) (the
minimum needed for a cross-submodule call from the new file).
render_text_scrollbar (render.rs) was private (not even pub(super)); needed
pub(super) to be reachable from tui::screens::preview.

Pure move — bodies and existing tests unchanged.

Part of #287
